arch_grpc_microservice/auth_service/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
bcrypt==4.2.0
pydantic==2.9.2